remove debug package
